Tinebase(Feature): make apply interceptable in simple dialogs
Reviewed-on: http://gerrit.tine20.com/customers/11533
Tested-by: Jenkins CI (http://ci.tine20.com/) <tine20-jenkins@metaways.de>

+++ tine20/Tinebase/js/widgets/dialog/Dialog.js
@@ -95,9 +95,12 @@ Tine.Tinebase.dialog.Dialog = Ext.extend(Ext.FormPanel, {
     },
 
     onButtonApply: function() {
-        this.fireEvent.apply(this, ['apply'].concat(this.getEventData('apply')));
-        this.purgeListeners();
-        this.window.close();
+        var eventData = this.getEventData('apply');
+        if (this.fireEvent.apply(this, ['beforeapply'].concat(eventData)) !== false) {
+            this.fireEvent.apply(this, ['apply'].concat(eventData));
+            this.purgeListeners();
+            this.window.close();
+        }
     },
 
     getCanonicalPathSegment: function () {
